stop automatically stripping Re: bits from message subjects when indexing
Rationale: when thread-by-subject is on, whether the subject is a reply or not
is actually used by threading to pick a root message (Thread#each). Stripping
the Re: header then gives incorrect behavior until the message has been loaded
from the source. Currently this isn't really a problem because threading isn't
called until all messages have been loaded from source, but in the future we
might optimize that a bit, at in the present, it gives weird results when
debugging.