maintain a more reasonable thread structure during threading
two new invariants: every Thread should have at least one Container
tree, and every Container tree should have at least one Message.
this is also a bugfix for threadset#remove_id, but the current use
of that function is a bug in and of itself, which I will fix in
the next commit. At any rate, this is better code, presumably.