add the ability to joining threads forcefully to ThreadSet
ThreadSet#join_threads now takes any number of threads (which must be part
of the threadset) heuristically either picks one message to be the root
(if all messages but one have a "Re:" on the subject line), or creates
a fake root. It updates the message object and marks it dirty so that a
subsequent save will preserve the new structure.