return sources in id-order, with non-archived sources first
To speed up the user's view of polling, deal with sources that affect
the inbox before (for example) large mailing lists. It seemed tidier to
implement this in Index rather than PollManager.

I changed "@sources.values" to "sources" in some other methods for
clarity, which I think in this case outweighs lost efficiency.