bugfix: hidden threads don't actually reappear when asked to unhide
In ThreadIndexMode, the logic was a little wrong. @hidden_threads isn't
keyed by Messages but by Threads, and for incoming messages that may
belong to a thread, it's necessary to find the corresponding thread first.

This was resulting in threads that had been archived in inbox-mode not
reappearing when new messages that belonged to them arrived.