allow user to make a gem with a specific version number
If one want to make a gem from a git version of sup, he can also
want to define a version number (e.g. in the form <previous public
release>.1, so that when a new version is released the gem will be
updated nonetheless). With this patch the version can be defined with
the REL environment variable (as in ferret Rakefile)