handle message "updates" by merging labels and overwriting other fields
Later versions of the same message typically happen because the original
was sent to a mailing list. This change merges in labels from later messages
and allows them to overwrite the other message fields (which was the
original behavior which this branch has broken).

This means that messages sent to mailing lists and then received back from
them preserve subject line rewriting and any auto-label application.