reload label list on focus
This ensures the unread count for each label is correct.

I often read my list mail from the label list. When I close the thread index
and go back to the label list the unread message count is wrong. This bugs me.

I'm sure a less brutal way of doing this is possible with the UpdateManager
but that seems complicated and therefore prone to errors.