maildir speedups
These changes improve the speed of maildir scanning by caching the mtime of the
relevant (cur, new) subdirectories and skipping the poll if the cached mtime
indicates that there are no new files to look at.  We also save a syscall when
creating an id for a maildir entry by calling stat and using its properties
instead of File.mtime and File.size (which are wrappers around a stat call).

Currently, mtimes are cached across session in sources.yaml.  We don't cache
the message id -> filename mapping (yet?) though, so this doesn't currently buy
much.