use offscreen lines for textfield overflow
In the To:, Cc:, Bcc:, etc., prompts, the TextField was initializing
an NCurses form with zero offscreen lines.  So when the user typed too
many addresses at the prompt the cursor wrapped around and everything
they had typed was clobbered.  This patch initializes the form object
to have 256 offscreen lines (I had no rationale for 256 except that it
is a nice, big power of two).  Now the user can type plenty of names
and the wrap-around works as the user would expect.