keybindings: ; -> buffer-list-mode, b, B, +
';' is mapped to buffer-list-mode. I really like this, because
then ;, j, k and enter are all you need to quickly swap between buffers.

Apply-to-tagged is now '+', from ';'. Sorry. But this also leaves
room for ':' as a future buffer-chooser device, e.g. by name, with tab
completion.

'B' is reenabled to mean swap-buffer-backwards.