put labels before subject in thread index view
This patch is probably controversial, and I expect it
to be rejected.  But I really like the way Gmail puts
the labels before the subject, and I've duplicated that here.
It helps out at work, where subject lines tend to be very
long, pushing the labels past the right edge of the window.