prevent "year too big to marshal" crashes
Xapian::Document#entry= performs a Marshal.dump which was throwing a
"year too big to marshal" when adding a new message.  I just
truncate_date'd the message's date when the entry hash is created in
sync_message and this problem went away for me.

It was a spam message, as the MIN_DATE/MAX_DATE comments hinted.