Add global keybinding to clear all hooks
When developing/tweaking hooks, it's inconvenient to use the console
mode to run the clear hooks command.  Add a globally available key
binding ('H') to trigger HookManager.clear.  As this is now more
exposed in the UI, make HookManager flash a notice to the
BufferManager indicating that it cleared the hooks.

Signed-off-by: Ben Walton <bwalton@artsci.utoronto.ca>