Allow toggle on Source.usual and Source.archived
Allow sources to be more malleable by allowing for the modification of
the usual and archived flags.  This is a first step toward patterned
(meta) sources.  Initially, the patterning will be handled by one of
the poll hooks, which will use the usual toggle when a 'new' source is
added.  Later on, this may be handled internally for sources with a
meta = true attribute (or something similar).

Signed-off-by: Ben Walton <bwalton@artsci.utoronto.ca>