Bugfix: Set background color for starred column
For the sake of consistency, Sup should always use the starred_color for the
'starred' column in the thread-index-mode, whether there's a star or a blank.

Rationale: With this patch, I'm able to set the background color of the
thread-index to bg:7 without having to change my term's background color.
Without the patch, the background color shines through in the starred column.