Jump and open next/previous message with C-n and C-p
Added functionality and key bindings for thread-view-mode where you can jump
_and_ open the next message. If a message has been opened by the jump it
will be closed if you continue to jump (in either direction). This is
saved in :toggled_state in the layout.

I often want to just jump to the next message to be able to skim through
a whole thread, if there was a key binding; jump and then hit enter
would work - but this patch gives me exactly what I want.

I'm not entirely happy with the key bindings, but there are not that many
left.

Signed-off-by: Gaute Hope <eg@gaute.vetsj.com>