mbox: fix date format
The date format used in all (known/common) variations is that of Unix/POSIX
asctime(), not the RFC 822 / RFC 2822 Date: header format.

Using an incorrect date format causes at least mutt, mb2md and Python
UnixMailbox (but not PortableUnixMailbox) to either reject the mailbox or show
it as empty.

References:
http://homepage.ntlworld.com./jonathan.deboynepollard/FGA/mail-mbox-formats.html
http://www.qmail.org/man/man5/mbox.html

Signed-off-by: Sascha Silbe <sascha-pgp@silbe.org>