Bugfix: mime_encode: Encode _ as =5F instead of =96
This line of code was introduced in 0d45aa38, probably by mistake. The ASCII
code in decimal for '_' (underscore) is 95 (the code says 96, but probably
another mistake), but as mime encoding is hexadecimal, we need =5F instead.

An example of this encoding problem can be seen in one of my earlier patches
with message-id <1287565270-sup-1161@midna.zekjur.net>.