Re-run the crypto selector hook after message changes.
The crypto-selector hook will be re-run after the message is edited,
or after the recipients change due to changing the "Reply to"
selector in reply-mode. So when the recipients change, the default
selector will also change.

However, if the user has manually changed the crypto selector then the
hook will not be re-run, so as to avoid over-riding the user's choice.