poll: Properly refresh thread-index-mode when a message has changed its source location
There are two cases when a message is detected as "updated", whether
its flags has changed or its location pathname has changed, eg. when
sending a message sup generates a Maildir pathname that is then
modified by offlineimap. It would be detected as deleting a message
and adding another one although it concern the same message, in that
case locations should be updated and its thread-index-mode entry
should be refreshed.