Check whether the "sync_back_to_maildir" option has been changed from false to true suggest the user to run sup-sync-back-to-maildir before continuing if needed
The check occurs when starting the Redwood module so it affects every
executables loading it since some of them can modify the index (sup,
sup-sync, etc.).

There is a new parameter called "bypass_sync_check" to the
Redwood::start method to bypass this check, it is useful for scripts
such as sup-sync-back-to-maildir itself!

It works by checking the presence of a flag file in the configuration
directory. The file is written when the synchronization is performed
and deleted when the "sync_back_to_maildir" option is set to false.