Ensure default options are set
When a new option is added (previous to this commit) then if it is not
in your config.yaml it would default to false, as sup would just load
your config file and not then check default values.

Now the default values (apart from accounts) are set and are merged with
the values from your file (with the file taking precedence of course). So
any new values added will have the intended default.