Allow user to permanantly skip syncback
Sup will warn you each time it starts if maildir syncback was not
performed and ask if you really want to continue. If you just started
using sup, this can be annoying, since calling the syncback script is an
unnecessary hassle.

This adds a second question to turn off the startup nag.