Merge #85: Check if GPG end marker exists
Picked up by @manizzle.

"lines.between() does not check if the end marker actually exists, but
later it is assumed the marker exists. This change introduces a check
for the end marker and doesn't decrypt messages in which the end marker
is missing. This is a fix for http://rubyforge.org/pipermail/sup-talk/2010-August/004209.html"

Source:
http://permalink.gmane.org/gmane.mail.sup.devel/641