Track change in syncback flag per source, warn user
Check per source if the sync_back flag changed from false to true. If so,
warn the user. sup-sync-back-maildir now stores the URIs of all sources that
have been synced by it in the file sync-back-ok, which we already write as a
status flag.

We need an extra verification step during startup, since index and sources
need to be initialized. It might be better to load the sources separately,
though, just to keep things nicely separated.