t_v_m: dont escape filename when saving attachments
Extra Shellwords.escaping was introduced in 3106f6d where in reality
only functionality for saving with default filenames should be added.

The Shellwords escaping is removed. The filename is not used in any
command, and any program or command out-of-sup needs to account for
correct filename. The filename may potentially contain shell commands,
the file will be saved with the filename indicated by the SENDER of the
email.

Fixes #277.