completion-mode: correctly set completion character
Auto completion is case-insensitive. Calculation of the completion
character must be case-insensitive as well.

Otherwise, if you have two labels "bar" and "Baz", and you type

    Lba<Tab>

the letters 'b' and 'B' are marked, when really 'r' and 'z' should be
marked. The same goes for contacts with an initial segment that differs
only in case.