Pass the message object to the "reply-to" hook.
This permits the hook to choose a different default mode for different
mailing lists.

Some mailing lists are mostly for discussion among the list members,
while others are used as points of contact for non-list-members to
mail a development team. If you're active on both kinds of mailing
list, you might well want to default to reply-to-list for the former
but reply-to-all for the latter; letting the "reply-to" hook see the
whole message allows its decision to depend on message.list_address.