Add a hook to filter text chunks
Currently, it's only possible to replace non-text chunks to text.
There is no way to change the content of an email, or control
whether a quote block is expanded or collapsed by default.

The direct motivation of adding this filter is because one of my email
providers wraps all URLs to the "urldefense" ones. For example, a plain
text hg patch containing the following lines:

  # Available At http://mercurial-scm.org/repo/users/marmoute/mercurial/
  # hg pull http://mercurial-scm.org/repo/users/marmoute/mercurial/ -r 1e21cde5

will be rewritten as:

  # Available At https://urldefense.proofpoint.com/v2/url?(275 chars omitted)
  # hg pull https://urldefense.proofpoint.com/(275 chars omitted) -r 1e21cde5

which actually stops the "hg pull" command from working properly.

Other features like being able to expand a small quote chunk and removing
boring signature chunks added by a mailing list are nice-to-have.