don't apply shell escaping to attachment filenames when saving
When we're saving attachments to disk, there is no shell involved, so
there is no need to escape spaces or other shell metacharacters. The
only character we need to escape is the directory separator.

Based on a patch contributed by
Felix Van der Jeugt <felix.vanderjeugt@gmail.com>
in PR #528.

Fixes #505.