fix charset handling for quotable attachments
This patch fixes two related issues when displaying text/plain
attachments, or attachments which have been converted to text by the
mime-decode hook.

When the attachment was text/plain and it declared a charset, we were
trying to transcode it back *into* its declared charset which is
backwards and doesn't work. Probably nobody ever noticed this because
text/plain attachments in non-ASCII charsets would be rare.

Instead, in that case we should be using String#force_encoding to tell
the string what encoding it is really in (or supposed to be in) before
we manipulate it any further.

On the other hand, when we get a string back from the mime-decode hook
we *don't* want to apply the original charset to it. There is not
necessarily any relationship between the input bytes and what is
produced by the hook.

For a specific example, Outlook sends messages with
Content-Type: text/html; charset=us-ascii. The mime-decode hook may
run that through elinks, which will always produce output in the user's
locale. If we reinterpret the hook's output as ASCII, any non-ASCII
characters coming from elinks will be ruined.