use Unicode::DisplayWidth to handle basic emoji
This fixes display glitches which can occur when the sender name or
subject line contain emoji.

We intentionally avoid using the unicode-display_width gem's optional
emoji support, because that turns out to be prohibitively slow for these
methods which are in the hot path of rendering the message list. With
this patch we will still count basic emoji characters themselves as
double-width. As far as I can tell, the only consequence of avoiding the
extra emoji support is that we will *overestimate* the length of emoji
ZWJ sequences. That's not the end of the world, since we are just
truncating strings here. And many terminal emulators still can't render
emoji ZWJ sequences correctly anyway.

The Unicode::DisplayWidth.of method is careful to always return
a value greater than zero so we can simplify the code on our side
a bit too.