don't attempt to fix bad encoding in RFC2047 words
If we find an RFC2047-encoded header word with an unrecognised charset
name or invalid bytes, let's just display it as is without decoding.

This is one of the options recommended in the RFC for MUAs to handle
words which cannot be decoded successfully, and it's more likely to
produce something readable than forcibly reinterpreting the string as
UTF-8, which is what our String#transcode method was doing.