sup-add: avoid URI.escape
This method was also removed in Ruby 3.0.

We can't use the new Source.encode_path_for_uri helper method here,
because sup-add expects a complete URI, not just a path. That method
will wrongly escape the colon in the URI scheme.

The simplest fix is to just remove the URI encoding and parsing here
entirely, and leave that until later when sup loads the sources.

Fixes #588.