use YAML::unsafe_load on Ruby 3.1+
In Ruby 3.1, Psych changed behaviour to make the YAML::load method be an
alias for safe_load, which has a restrictive whitelist of classes which
can be instantiated by the YAML. A new method YAML::unsafe_load gives
the old behaviour, allowing any class to be instantiated.

Sup only parses YAML from trusted sources (the user's config files) so
use YAML::unsafe_load when it's available, to get the old behaviour.