restore monkey-patching of RMail::Serialize.calculate_boundaries
The monkey-patch was not applying due to module load ordering. It seems
to still be needed in order to avoid superfluous MIME-Version headers on
nested signed or encrypted parts, although such headers are likely not
harmful.

Includes test cases which attempt to reproduce the original behaviour
and fix that commit 07f0be01 probably intended.