add nix-shell expressions
One for each version of Ruby back to Ruby 2.4, so that I can run sup and
its tests under all these versions reproducibly.

These should also work on MacOS, so they can be used for CI and they
will give me a greater chance of reproducing and fixing the problems
there, rather than using all that Homebrew stuff that I am clueless
about.

The separate copy of the Gemfile is needed because bundix cannot
understand 'gemspec' calls in Gemfiles, it needs the dependencies to be
spelled out explicitly. So this Gemfile is just a copy-paste of the same
dependency declarations from sup.gemspec.