fix signing key selection for sign_and_encrypt
The GPGME::Crypto#encrypt() method looks for :signers in its options
hash, but Sup was passing :signer which has no effect. As a result, the
mail would always be signed with the user's default signing key.

Keep the :signer misspelling on the Sup side, to avoid breaking anyone's
existing 'gpg-options' hooks, but pass :signers to GPGME.

Fixes #522.