parse drafts as individual messages, not mboxes
As far as I can see they were always being written with one message per
file and no From envelope (maildir style) so there is no reason to try
iterating the draft as if it were an mbox.

Fixes the pathological case of trying to load an empty file as a draft.