rescue TypeError when sending mails
If the user has not set the "sendmail" key in their account config, we
will try to invoke Popen on nil, which is a TypeError.

Showing this TypeError to the user in a flash message is not very nice,
but it's still better than crashing and losing their draft.

Relates to #524.